requirements.txt
torch==2.9.1
torchvision==0.24.1
transformers==4.57.3
diffusers==0.36.0
safetensors==0.7.0
numpy==2.4.0
pillow==12.1.0
requests==2.32.5
scipy==1.16.3